agents/notion_content_agent/requirements.txt
notion-client>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
rich>=13.0.0
